fix(ci): move useConversationDrop into fileUpload/helpers to satisfy caps
The prettier reformat in 177617ea grew ThreadConversation.tsx to 504
lines (>500), failing both the Structure and Frontend (eslint max-lines)
CI gates. Extracting the OS-file drag-and-drop hook fixes the line cap,
but a standalone fileUpload/useConversationDrop.ts pushed the folder to 9
entries (>8). Folding the hook into the existing fileUpload/helpers.ts
(the fileUpload catch-all — types + pure parsers) satisfies both caps at
once: ThreadConversation.tsx drops to 412 lines and fileUpload/ stays at
8 entries. Pure relocation, no behaviour change. Structure, ts-lints and
mobile-mirror all green locally.
